Decomposed types to separate files.
