+++ src/components/Cart/CartBillingForm/CartBillingForm.module.scss
@@ -35,10 +35,10 @@
   gap: 5px;
 
   text-wrap: wrap;
+}
 
-  & > a {
-    text-decoration: underline;
-  }
+.agreement_links {
+  text-decoration: underline;
 }
 
 .marketing {
