+++ src/types/form/validationsTypes.ts
@@ -1,0 +1,1 @@
+export type Validations = Record<string, string | number | boolean>;
