+++ src/types/index.ts
@@ -3,3 +3,5 @@ export * from './store';
 export * from './productFilter';
 export * from './pagination/pagination';
 export * from './address/addressTypes';
+export * from './form/formTypes';
+export * from './form/validationsTypes';
